+++ src/interventions/scheduled.jl
@@ -116,6 +116,7 @@ end
 end
 
 required_fields(s::Scheduled) = required_fields(s.intervention)
+_unwrap_scheduled(s::Scheduled) = _unwrap_scheduled(s.intervention)
 intervention_time(s::Scheduled, ind::Individual) = intervention_time(s.intervention, ind)
 reset!(s::Scheduled, ind::Individual) = reset!(s.intervention, ind)
 # On the continuous-time models a Scheduled removes a case when the wrapped